pyproject.toml
[tool.poetry]
name = "boosting_service_django"
version = "0.1.0"
description = "Boosting web application with using django"
authors = ["Nikita"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.8.1"
babel = "^2.14.0"
build = "^1.2.1"
cachecontrol = "^0.14.0"
certifi = "2024.2.2"
cffi = "^1.16.0"
charset-normalizer = "^3.3.2"
cleo = "^2.1.0"
colorama = "^0.4.6"
crashtest = "^0.4.1"
cryptography = "^42.0.5"
defusedxml = "^0.7.1"
distlib = "^0.3.8"
django = "4.2"
django-allauth = "0.61.1"
django-money = "3.4.1"
django-timezone-field = "6.1.0"
dulwich = "^0.21.7"
fastjsonschema = "^2.19.1"
filelock = "^3.13.4"
idna = "^3.7"
importlib-metadata = "^7.1.0"
installer = "^0.7.0"
jaraco-classes = "^3.4.0"
keyring = "24.3.1"
more-itertools = "^10.2.0"
msgpack = "^1.0.8"
oauthlib = "^3.2.2"
packaging = "^24.0"
pexpect = "^4.9.0"
pillow = "10.2.0"
pip = "23.3.2"
pkginfo = "1.10.0"
platformdirs = "4.2.0"
poetry = "1.8.2"
poetry-core = "1.9.0"
poetry-plugin-export = "1.7.1"
ptyprocess = "0.7.0"
py-moneyed = "^3.0"
pycparser = "^2.22"
pyjwt = "^2.8.0"
pyproject-hooks = "^1.0.0"
python3-openid = "^3.2.0"
pywin32-ctypes = "^0.2.2"
rapidfuzz = "^3.8.1"
requests-oauthlib = "^2.0.0"
requests-toolbelt = "^1.0.0"
setuptools = "^69.5.1"
shellingham = "^1.5.4"
sqlparse = "^0.5.0"
tomlkit = "^0.12.4"
trove-classifiers = "2024.4.10"
typing-extensions = "^4.11.0"
tzdata = "^2024.1"
urllib3 = "^2.2.1"
virtualenv = "20.25.3"
wheel = "^0.43.0"
zipp = "^3.18.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
